refactor: widen the PRNG callback from int16_t to int
The C library states a generator's range with RAND_MAX and returns a
plain int; this callback stated it twice, once in KUSOKURAE_RAND_MAX and
once in a return type that permits a different set of values. The two
disagreed exactly on the negative numbers, which is where the overrun
fixed in the previous commit lived. int leaves the range to
KUSOKURAE_RAND_MAX alone, lets a generator covering [0, 65535] say so
instead of silently wrapping half its values negative, and leaves room to
raise the ceiling later without touching the signature.

Widening it immediately surfaced the mirror-image defect. A draw above
KUSOKURAE_RAND_MAX was not expressible before, and it turns out to be
just as damaging as a negative one: it fails comparisons it should have
passed, so a hand ends up short, with untouched slots still in it. The
new prngcontract case caught it on the first run -- 3 to 4 cards missing
per deal.

sample() now bounds the selection on both sides: rwanted >= rcount takes
the card, rwanted == 0 never does. Together they make "exactly wanted
items are chosen" hold for any generator at all, instead of only for one
that honours the contract. Both bounds are no-ops for a conforming
generator, verified by hashing every card of 40000 deals across three-
and four-player games against fixed seeds: identical before and after.

sm.h explains the choice of int and says what an out-of-range value now
does in each direction.

+++ sm/benchmarks/dealbench.c
@@ -60,7 +60,7 @@ static int deal_selection(void *state) {
         kusokurae_card_t *psrc = pass[k].src, *pdst = pass[k].dst, *prej = pass[k].rej;
         size_t rc = pass[k].count, rw = pass[k].wanted;
         while (rc > 0) {
-            int16_t dice = ms_rand(state);
+            int dice = ms_rand(state);
             draws++;
             if (dice < (int64_t)(KUSOKURAE_RAND_MAX + 1ULL) * rw / rc) {
                 *pdst++ = *psrc;
