+++ sm/prng_bridge.go
@@ -8,19 +8,19 @@ package sm
 /*
 #include "sm_internal.h"
 
-extern void goRandom(int16_t *);
+extern void goRandom(int *);
 
 // Mirrors the bridge sm.go installs in init(): the C engine calls back into Go
 // for every random number it needs.
-static int16_t test_cgo_random(void *state) {
-	int16_t ret;
+static int test_cgo_random(void *state) {
+	int ret;
 	goRandom(&ret);
 	return ret;
 }
 
 // Drives that bridge from the C side, i.e. the exact Go -> C -> Go round trip
 // the engine performs per dice roll.
-static int16_t call_bridge_from_c(void) {
+static int call_bridge_from_c(void) {
 	return test_cgo_random(NULL);
 }
 
@@ -43,20 +43,20 @@ const cPRNGMax = C.KUSOKURAE_RAND_MAX
 
 // goRandomValue calls the Go half of the bridge directly, without crossing any
 // language boundary.
-func goRandomValue() int16 {
-	var out C.int16_t
+func goRandomValue() int {
+	var out C.int
 	goRandom(&out)
-	return int16(out)
+	return int(out)
 }
 
 // nativeRandom calls the engine's own PRNG (one Go -> C crossing).
-func nativeRandom(state *int32) int16 {
-	return int16(C.ms_rand(unsafe.Pointer(state)))
+func nativeRandom(state *int32) int {
+	return int(C.ms_rand(unsafe.Pointer(state)))
 }
 
 // randomViaC goes Go -> C -> Go, the full round trip the engine pays for.
-func randomViaC() int16 {
-	return int16(C.call_bridge_from_c())
+func randomViaC() int {
+	return int(C.call_bridge_from_c())
 }
 
 // useNativePRNG makes the engine use its built-in PRNG.
